v9

+++ Assets/Scripts/Player/PlayerWallet.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [RequireComponent(typeof(BoxCollider2D))]
 public class PlayerWallet : MonoBehaviour
 {
-    [SerializeField] private int _score = 0;
+    [SerializeField] private float _score = 0;
 
     private void OnTriggerEnter2D(Collider2D other)
     {
