+++ Assets/Scripts/Projectiles/Damager.cs
@@ -2,10 +2,10 @@ using UnityEngine;
 
 public class Damager : MonoBehaviour
 {
-    [field: SerializeField] public int Damage { get; private set; }
+    [field: SerializeField] public float Damage { get; private set; }
     [field: SerializeField] public bool IsAlly { get; private set; }
 
-    public void SetParameters(int damage, bool isAlly)
+    public void SetParameters(float damage, bool isAlly)
     {
         Damage = damage;
         IsAlly = isAlly;
