+++ Assets/Scripts/Surroundings/Coin.cs
@@ -5,7 +5,7 @@ public class Coin : MonoBehaviour
 {
     public event Action<Coin> IsCollected;
 
-    [field: SerializeField] public int Value { get; private set; } = 10;
+    [field: SerializeField] public float Value { get; private set; } = 10;
 
     public void MarkAsCollected()
     {
