+++ Assets/Scripts/Surroundings/Medpack.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Medpack : MonoBehaviour
 {
-    [field: SerializeField] public int HealthAmount { get; private set; } = 2;
+    [field: SerializeField] public float HealthAmount { get; private set; } = 2;
 
     public void CommandDestruction()
     {
